fixed the renaming from backends to queues
